refactor: restructure collectAdmonitions command and related utilities
