+++ src/command/command/index.mts
@@ -1,0 +1,3 @@
+import collectAdmonitions from "./collectAdmonitions/index.mjs";
+
+export { collectAdmonitions };
